requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Caching==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
